+++ radioactive/__main__.py
@@ -4,13 +4,13 @@ import signal
 import sys
 from time import sleep
 
+from pick import pick
 from rich import print
 from rich.console import Console
 from rich.panel import Panel
 from rich.table import Table
 from rich.text import Text
 from zenlog import log
-from pick import pick
 
 from radioactive.alias import Alias
 from radioactive.app import App
@@ -19,16 +19,6 @@ from radioactive.handler import Handler
 from radioactive.help import show_help
 from radioactive.last_station import Last_station
 from radioactive.player import Player, kill_background_ffplays
-
-
-# using sentry to gather unhandled errors at production and will be removed on next major update.
-# I respect your concerns but need this to improve radioactive.
-import sentry_sdk
-
-sentry_sdk.init(
-    dsn="https://e3c430f3b03f49b6bd9e9d61e7b3dc37@o615507.ingest.sentry.io/5749950",
-    traces_sample_rate=1.0,
-)
 
 RED_COLOR = "\033[91m"
 END_COLOR = "\033[0m"
@@ -44,7 +34,7 @@ def main():
     app = App()
     args = parser.parse()
 
-    ############ all the args ############
+    # ----------------- all the args ------------- #
     show_help_table = args.help
     station_name = args.station_name
     station_uuid = args.station_uuid
@@ -61,9 +51,9 @@ def main():
     show_favorite_list = args.show_favorite_list
     flush_fav_list = args.flush
     kill_ffplays = args.kill_ffplays
-    ########################################
 
     VERSION = app.get_version()
+
     if args.version:
         log.info("RADIO-ACTIVE : version {}".format(VERSION))
         sys.exit(0)
@@ -77,32 +67,31 @@ def main():
     else:
         log.warning("Correct log levels are: error,warning,info(default),debug")
 
-    handler = Handler()
-
     mode_of_search = ""
     direct_play = False
     direct_play_url = ""
     skip_saving_current_station = False
-    is_alias = False
 
-    ################################# RICH ##################
+    # --------------- RICH ------------------- #
     console = Console()
 
     welcome = Panel(
         """
         :radio: Play any radios around the globe right from this Terminal [yellow][blink]:zap:[/blink][/yellow]!
         :smile: Author: Dipankar Pal
-        :question: Type '--help' for more details on available commands.
-        :bug: Visit https://github.com/deep5050/radio-active to submit issues
+        :question: Type '--help' for more details on available commands
+        :bug: Visit: https://github.com/deep5050/radio-active to submit issues
         :star: Show some love by starring the project on GitHub [red][blink]:heart:[/blink][/red]
         :dollar: You can donate me at https://deep5050.github.io/payme/
         :x: Press Ctrl+C to quit
         """,
-        title="[b][rgb(250,0,0)]RADIO[rgb(0,255,0)]ACTIVE[/b]",
+        title="[b]RADIOACTIVE[/b]",
         width=85,
     )
+
     print(welcome)
 
+    handler = Handler()
     alias = Alias()
     alias.generate_map()
     last_station = Last_station()
@@ -111,7 +100,7 @@ def main():
         update_msg = (
             "\t[blink]An update available, run [green][italic]pip install radio-active=="
             + app.get_remote_version()
-            + "[/italic][/green][/blink]\n See the changes: https://github.com/deep5050/radio-active/blob/main/CHANGELOG.md"
+            + "[/italic][/green][/blink]\nSee the changes: https://github.com/deep5050/radio-active/blob/main/CHANGELOG.md"
         )
         update_panel = Panel(
             update_msg,
@@ -144,6 +133,7 @@ def main():
             for entry in alias.alias_map:
                 table.add_row(entry["name"], entry["uuid_or_url"])
             console.print(table)
+            log.info(f"Your favorite stations are saved in {alias.alias_path}")
         else:
             log.info("You have no favorite station list")
         sys.exit(0)
@@ -160,20 +150,19 @@ def main():
         sys.exit(0)
 
     # ------------------ discover ------------------ #
-    _limit = int(limit) if limit else 100
+    _limit_ = int(limit) if limit else 100
 
     if discover_country_code:
-        # search for stations in your country
-        handler.discover_by_country(discover_country_code, _limit)
+        handler.discover_by_country(discover_country_code, _limit_)
 
     if discover_state:
-        handler.discover_by_state(discover_state, _limit)
+        handler.discover_by_state(discover_state, _limit_)
 
     if discover_language:
-        handler.discover_by_language(discover_language, _limit)
+        handler.discover_by_language(discover_language, _limit_)
 
     if discover_tag:
-        handler.discover_by_tag(discover_tag, _limit)
+        handler.discover_by_tag(discover_tag, _limit_)
 
     # -------------------- NOTHING PROVIDED --------------------- #
     # if neither of --station and --uuid provided , look in last_station file
@@ -184,10 +173,11 @@ def main():
 
         try:
             last_station_info = last_station.get_info()
-        except:
+        except Exception as e:
+            log.debug("Error: {}".format(e))
             # no last station??
             pass
-        # print(last_station_info)
+
         log.info("You can search for a station on internet using the --station option")
         title = "Please select a station from your favorite list:"
         station_selection_names = []
@@ -200,7 +190,8 @@ def main():
             )
             try:
                 station_selection_urls.append(last_station_info["stationuuid"])
-            except:
+            except Exception as e:
+                log.debug("Error: {}".format(e))
                 station_selection_urls.append(last_station_info["uuid_or_url"])
 
         fav_stations = alias.alias_map
@@ -221,7 +212,9 @@ def main():
 
         # check if there is direct URL or just UUID
         station_option_url = station_selection_urls[index]
-        station_name = station_selection_names[index]
+        station_name = station_selection_names[index].replace(
+            "(last played station)", ""
+        )
 
         if station_option_url.find("://") != -1:
             # set direct play to TRUE
@@ -231,43 +224,13 @@ def main():
             # UUID
             station_uuid = station_option_url
 
-    ##################################
-
-    # try:
-    #     if last_station_info["alias"]:
-    #         is_alias = True
-    # except:
-    #     pass
-
-    # if is_alias:
-    #     alias.found = True  # save on last_play as an alias too!
-    #     # last station was an alias, don't save it again
-    #     skip_saving_current_station = True
-    #     station_uuid_or_url = last_station_info["uuid_or_url"]
-    #     # here we are setting the name but will not be used for API call
-    #     station_name = last_station_info["name"]
-    #     if station_uuid_or_url.find("://") != -1:
-    #         # Its a URL
-    #         log.debug(
-    #             "Last station was an alias and contains a URL, Direct play set to True"
-    #         )
-    #         direct_play = True
-    #         direct_play_url = station_uuid_or_url
-    #         log.info("Current station: {}".format(
-    #             last_station_info["name"]))
-    #     else:
-    #         # an UUID
-    #         station_uuid = last_station_info["uuid_or_url"]
-    # else:
-    #     # was not an alias
-    #     station_uuid = last_station_info["stationuuid"]
-    ############################################
-
     # --------------------ONLY UUID PROVIDED --------------------- #
     # if --uuid provided call directly
     result = None
     if station_uuid is not None:
         mode_of_search = "uuid"
+        log.debug("increased click count for: {}".format(station_uuid))
+        handler.vote_for_uuid(station_uuid)
 
     # ------------------- ONLY STATION PROVIDED ------------------ #
 
@@ -291,8 +254,10 @@ def main():
                     log.debug("Entry contains a UUID")
                     # mode_of_search = "uuid"
                     station_uuid = result["uuid_or_url"]  # its a UUID
+                    handler.vote_for_uuid(station_uuid)
 
-            except:
+            except Exception as e:
+                log.debug("Error: {}".format(e))
                 log.warning("Station found in favorite list but seems to be invalid")
                 log.warning("Looking on the web instead")
                 alias.found = False
@@ -340,19 +305,25 @@ def main():
 
     # TODO: handle error when favouring last played (aliased) station (BUG) (LOW PRIORITY)
     if add_to_favorite:
-        alias.add_entry(add_to_favorite, handler.target_station["url"])
+        try:
+            alias.add_entry(add_to_favorite, handler.target_station["url"])
+        except Exception as e:
+            log.debug("Error: {}".format(e))
+            log.error("Could not add to favorite. Already in list?")
 
     curr_station_name = station_name
 
     try:
-        # TODO fix this. when aliasing a station with an existing name curr_station_name is being None
+        # TODO fix this. when aliasing a station with an existing name
+        # curr_station_name is being None
         panel_station_name = Text(curr_station_name, justify="center")
 
         station_panel = Panel(
             panel_station_name, title="[blink]:radio:[/blink]", width=85
         )
         console.print(station_panel)
-    except:
+    except Exception as e:
+        log.debug("Error: {}".format(e))
         # TODO handle exception
         pass
 
@@ -362,7 +333,8 @@ def main():
     else:
         try:
             signal.pause()
-        except:
+        except Exception as e:
+            log.debug("Error: {}".format(e))
             pass